# perf: use mimalloc for musl target