@hardfist 2024-03-30
!bench