@h-a-n-a 2024-04-01
Seems like we can use mimalloc for everything. https://github.com/napi-rs/node-rs/blob/70d7fa72262c6e547950b30daa2d03583a1b04bd/crates/alloc/Cargo.toml