@xc2 2024-04-01
> Seems like we can use mimalloc for everything. https://github.com/napi-rs/node-rs/blob/70d7fa72262c6e547950b30daa2d03583a1b04bd/crates/alloc/Cargo.toml

yea it's good to make the allocator unified but i didn't make the change as i found it didn't help much on performance for glibc linked binary. 

and i need more effort to check if these change below works well. let me check if i got time later.
- mimalloc-rust -> mimalloc 
- tikv-jemalloc -> mimalloc